+++ src/components/Header/CloseButton.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
 import IconButton from '../Button/IconButton';
